reorganize folders and files
